Fix the handling of arguments to dt_predicate and dt_program
The trampoline code that all providers generate a custom version of
is called as a BPF program when a probe triggers.  It is passed a
probe specific BPF context (ctx).  The trampoline code itself is
tasked with setting up a DTrace context (dctx) and then executing
the following pseudo-code:

[if there is a predicate]
		rc = dt_predicate(ctx, dctx);
		if (rc == 0)
			goto out;
		rc = dt_program(ctx, dctx);
	out:
		return rc;

[if there is no predicate]
		rc = dt_program(ctx, dctx);
		return rc;

We know that the BPF context is passed to our trampoline program as
the first (and only) argument, and it is therefore available as %r1.

The trampolines were generating code that worked as follows:

		/* we know ctx is in %r1 */
		if (haspred) {
			%r6 = %r1;
			%r2 = &dctx;
			%r0 = dt_predicate(%r1, %r2)
			if ((%r0 == 0)
				goto out;
		}

		%r1 = %r6;
		%r2 = &dctx;
		%r0 = dt_program(%r1, %r2);
	out:
		return %r0;

The problem is that if there is no predicate, %r6 is never assigned
the value of %r1, and therefore we end up calling dt_program with an
invalid BPF context.  If there is a predicate, we end up determining
the address of dctx twice (which is not wrong but it is silly.

The new implementation implements the following:

		/* we know ctx is in %r1 */
		%r2 = &dctx;

		if (haspred) {
			%r6 = %r1;
			%r7 = %r2;
			%r0 = dt_predicate(%r1, %r2)
			if ((%r0 == 0)
				goto out;
			%r1 = %r6;
			%r2 = %r7;
		}

		%r0 = dt_program(%r1, %r2);
	out:
		return %r0;

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>