Fix register leak in trace()
The trace() implementation was incorrectly assuming that the value to be
written to the scratch buffer would be in %r0 (which is not guaranteed),
and it was not releasing the register that the value actually was stored
in.  THe code now uses the register assigned to the passed argument, and
ensures that it is released.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>