Reverse register allocation order and add explicit alloc
The allocation from %r0 through %r8 caused clobbering registers that
are needed for function calls (%r1 through %r5 for argument passing,
and %r0 for the return value).  By reversing the allocation order we
avoid this collisions (for now - we may need spilling later).

This patch also implements allocation of specific registers as a way
to ensure that e.g. registers to hold function call arguments are
available.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>