BPF asm macros do not wrap arguments in parentheses
For example, the definition of BPF_ALU64_REG in include/bpf_asm.h is:

    #define BPF_ALU64_REG(op, dst, src)     \
        ((struct bpf_insn) {                \
            .code = BPF_ALU64 | op | BPF_X, \
            .dst_reg = dst,                 \
            .src_reg = src,                 \
            .off = 0,                       \
            .imm = 0                        \
        })

When "op" is provided as an expression, the above macro can yield
unexpected results.  The following invocation illustrates this:

    # dtrace -Sen 'write:entry { trace(htonl(104)); }' > out.txt 2> err.txt
    # grep -A 1 "lsh  %r8" err.txt
    061 488: 6f 8 7 0000 00000000    lsh  %r8, %r7
    062 496: c0 8 7 0000 00000000    (illegal opcode)

This particular problem comes from libdtrace/dt_cg.c:

    21 #include <bpf_asm.h>

  1119 static void
  1120 dt_cg_typecast(...)
  1122 {

  1152     instr = BPF_ALU64_REG((dst->dn_flags & DT_NF_SIGNED) ?
  1153         BPF_ARSH : BPF_RSH, dst->dn_reg, reg);

Here the "op" is a ternary expression that causes the following macro
expansion:

    .code = BPF_ALU64 | (dst->dn_flags & DT_NF_SIGNED) \
        ? BPF_ARSH : BPF_RSH | BPF_X,

This is obviously not the intended effect.

Edit the bpf_asm.h file so that macros wrap their arguments with
parentheses when they use them.

Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>