Create 'mem' map as a BPF per-CPU array map
This patch adds a global variable to dtrace_hdl_t to track the largest
seen trace buffer record size for the programs being loaded.  Once all
programs have been compiled, this can be used to create the BPF per-CPU
array map that provides scratch memory for constructing a trace record.

The 'mem' BPF map contains a singleton element that is large enough to
construct any of the trace records that will be used by any of the
programs being loaded.  Since the map is created as a per-CPU map, each
CPU will get its own copy of the singleton element (key 0).

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>