Add logic to load BPF programs into the kernel
This patch introduces the code to load BPF program into the kernel.
Provider implementations are now required to provide a probe type
(BPF program type) that is to be used to load BPF programs.  In future
revisions, this may need to be finetuned more as providers might serve
probes that map to different BPF program types.

Right before we load BPF programs, we resolve relocations for BPF maps.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>