Create DIFO for BPF functions in an ELF object
This patch changes the way BPF functions are loaded from a BPF library.
WHere previously, functions and relocations were stored in global
variables, the new code associates specific function information with
the function identifier (as an opaque data item).  We also now introduce
'fake' function information for BPF maps.

Quite a bit of work is done because we need to deal with a few glitches
in the gcc BPF support (primarily the fact that all symbols have a zero
size).  We also need to split up the relocations so that we can associate
them with the function that they apply to.  This is necessary because
individual functions will be appended to the BPF program that uses them.
Each function will have its own DIFO, and therefore its own relocations.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>