Make dt_dis_difo() global
The dt_dis_difo() function is the actual implementation of the code
disassembler.  It was previously only used for providing disassembly
for a program compiled with DTrace.  With the use of pre-compiled BPF
code, there is a use case for calling it from the BPF library loader
code.  This symbol is being made available in preperation for that.
The goal is to add an option to request disassembly of loaded BPF
code for debugging purposes.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>