Use relocations in disassembler annotations where possible
Various instructions were already looking in the BPF relocation data
to try to add useful annotations to disassembler output.  This patch
provides a more generic support for relocation data in annotations.

It also adds local variable annotations to store instructions.  Note
that the inclusion of the trampoline code in the actual program code
can cause annotations to be emitted for local variable usage that is
not actually accurate because the trampoline is acting on a different
stack segment than the actual program (which which we have local
variable data).  The code cannot distinguish between the two different
stacks and therefore interprets stack offsets incorrectly.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>