Fix predicate return value passing
The code to generate predicates was not ensuring that the return value
was being passed in the %r0 register as required by BPF.  We need to
add a 'mov %r0, $rX' where %rX is the register where the code generator
stores the value that is computed in the predicate code.  This will
generally be %r8 but there is no need to count on that.

It is possible that code will be generated for the predicate that ends
with a sequence like:

	mov %r8, <value>
	mov %r0, %r8

This can obviously be optimized to simply be:

	mov %r0, <value>

We do not do that there (yet) because such sequences will be generated
in many more places during code genration, and it makes more sense to
implement this as part of an actual optimizer step rather than in every
case where it might occur.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>