Ensure that the return value of BPF map lookups is validated
When a value is retrieved from a BPF map, it is considered either a
valid map value of NULL.  BPF programs are required (by the verifier)
to validate that the return value is not NULL prior to using the
value.  We failed to do that in the prologue generated for compiled D
programs, resulting in a complaint from the verifier.

This patch also introduces a default label that will be placed at the
exit of the generated program.  This can be used by generated code to
jump right to the end, bypassing any further execution.  This is used
in the map lookup validation: if we were not able to retrieve a valid
value, we skip all further execution by jumping to the exit.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>