Fix invalid use of %r2 after function call (it got clobbered)
Registers %r0 through %r5 get clobbered by function calls, regardless
of whether the called function touches them or not.  This causes a
problem for the prologue implementation because it expected that it
could use %r2 after function call.

We now explicitly retrieve the value again (from the stack) and we
switch to using %r0 since we only use it as a temporary base pointer
in the next load instruction.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>