test: enable unittest/version/tst.1.0.d and add err.1.1.d
DTrace v2 should be able to pass unittest/version/tst.1.0.d and the
new err.1.1.d test, so they should be enabled in the testsuite.

The new err.1.1.d verifies the opposite case of tst.1.0.d.  In the
tst.1.0.d test we verify that identifiers that became a builtin at
version 1.1 can be used (e.g. as variables) if we run in version
1.0 mode.  The err.1.1.d test verifies that if we are in version
1,1 mode, those same identifiers are no longer available for our
use.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>