Give error for lack of representative probe
When a probe specification could not be resolved into a representative
probe (i.e. all probes that match the specification have the same
attributes so we can pick any one of them to use during compilation),
we would get a SEGV.

Issue an error if no representative probe can be found, and mark it as
a future feature.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>