test: aggs/tst.aggmod_full2.sh gives false PASSes
If dtrace fails to compile the D script, it will never open the output
file (output.txt) specified on the command line.  This was causing the
test script to believe that the test ran correctly.

While the test script had code in it to try to handle dtrace failures,
it incorrectly was using the return code of a multi-stage pipeline as
the return value of dtracer (the first command in the pipeline).  That
is wrong because the return code of a pipeline is the return code of the
last command in the pipeline.

This patch verifies whether output.txt exists.  If it does, dtrace got
past the compilation stage and will try to run the script (which may
still fail).  If it does not exist, the compilation failed.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>