test: ensure that the pre-release banner does not interfere with testing
The pre-release version of DTrace v2 outputs a banner message to inform
the user of the pre-release status of this tool.  Tests that verify
output were reporting failure because of the unexpected extra string
in the output.

This commit filters out the pre-release banner before any output
processing takes place.

Orabug: 30619056
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>