Build against libctf (from binutils) if available
CTF suypport can be provided by the libdtrace-ctf library (originally
part of the legacy DTrace implementation) or by the libctf that is
included in newer binutils versions.  The config detection mechanism
added to determine whether libctf is available provides a HAVE_LIBCTF
define to use in source code and also a HAVE_LIBCTF symbol that can be
tested for in makefiles.

This patch provides the changes in the DTrace v2 source tree to build
against libctf when it is available.  When it is not available, building
agaist libdtrace-ctf is attempted (and if that is not found a build
error will result).

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>