Ensure that all probes have a perf event ID
The implementation up until now has only supported retrieving the perf
event ID for tracepoints that are used for SDT and syscall probes.
FBT probes are a bit different because there are no perf events for
the kprobes that implement them until they are explicitly registered.

The new code for processing perf event data centralizes more of the
processing (especially the skipping of fields that we do not need),
and allows for providers not to care about argument information for
the probe/event.  This is used for FBT probes because those do not
provide argument information anyway.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>