Associate BPF programs with probes (as perf events)
Once a BPF program has been loaded successfully (i.e. the verifier has
accepted it), it needs to be associated (attached) to one or more
probes.  Each probe represents a perf event (identified by a unique
numeric id), and that event must be created by means of the
perf_event_open() system call.  The result (if successful) will be a
file descriptor that refers to the perf event for the probe.  This fd
can be used to attach the BPF program (identified by a file descriptor)
to the perf event.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>