Add BPF source code for precompiled BPF function library
Some functionality of the D language is implemented as BPF functions.
They are to be compiled during the build of the DTrace v2 application,
although that is not yet implemented in the build system.

The inclusion of the BPF source code required some restructuring of
include files in the source tree, specifically as it relates to the
dt_pt_regs structure and the DTrace BPF context (struct dt_bpf_context)
because they need to be available to both the compilation of the
DTrace library and the BPF function library.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>