Ensure all datatypes are available for compilation
The changes to make building against libctf as an alternative to
libdtrace-ctf caused a compilation failure if libdtrace-ctf was not
installed.  The cause was a dependency on some datatype declarations
defined by the libdtrace-ctf include files.  These declarations have
been moved to dtrace_types.h to ensure that they are available
regardless of which CTF library is used during compilation.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>