Add the precompiled BPF function library as a build component
The previously added BPF function library source code is being replaced
by the source code in the bpf/ directory.  It makes use of the new build
system support for compiling BPF code, thereby providing a better
integration with the DTrace build infrastructure.

(The build system changes for BPF code and this replacement for the
 previously monolithic bpf_dlib.c are inspired by a patch by Eugene Loh
 providing build time compilation of bpf_dlib.c.)

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>