ensure global and TLS variables map to keys starting at 0
Global and TLS variables are assigned an id from DIF_VAR_OTHER_UBASE
onward in order to distinguish them from built-in global variables.
Up until now the code generator was applying a base offset to the
variable ids for debugging purposes.

This patch adds support for distinguishing built-in variables from
user defined global variables.  All variables in a specific class
or scope should be mapped to keys that are assigned from 0 onward.

Built-in variables are loaded using dt_get_bvar.
Global variables are loaded using dt_get_gvar, and stored using
dt_set_gvar.
TLS variables are loaded using dt_get_tvar, and stored using
dt_set_tvar.

This patch also adjusts the disassembler to ensure that it is able to
retrieve the correct variable name (and scope).

The support for TLS variables is still left unimplemented at this
point and it is therefore encoded as if they are regular global
variables.  A correct implementation is forthcoming.

The dt_get_bvar() BPF function is a placeholder that returns -1 for
every builtin variable.  This patch also adds a tests for the entire
collection of available builtin variables, and are expected to fail
(for now) because none of the actual builtin variables have been
implemented yet.

Orabug: 31220516
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>