Implement perf event buffer management for DTrace
With the switch to using BPF as execution engine, DTrace must now also
use the perf event buffer mechanism to process trae data.  This patch
provides an implementation of the perf event buffer consumer.

The bulk of the buffer processing infrastructure is retained because
only the underlying mechanism is changing.  The biggest change is found
in the use of epoll_wait() to wait for data to become available in any
of the CPU buffers (bound by a timeout) instead of simply retrieving
buffer data at a preset rate.

TODO: Integration of the process death tracking with the new epoll based
      buffer polling implementation.  Rather than using a condition
      variable to signal a process notification, we can use an eventfd
      to communicate the fact that there are process notifications to
      process.

(Eugene Loh implemented a prototype of this patch based on earlier code
 that I had written.  This patch is based on my earlier work and builds
 on Eugene's work.)

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>