Convert the epid tracking code from discovery to explicit construction
In legacy DTrace (v1) the mapping from EPID to probe and related data
was constructed on the fly as tracing data was being processed.  When
an unknown EPID was seen, the DTrace kernel component was queried
(using ioctls) to obtain information about the enabling and probe.

Now that we manage all probes and enablings in userspace we already
know all the information about enablings and probes.  We can therefore
simply add them to the list of EPIDs for easy lookup when we start
processing tracing data.

The epid management code in dt_map.c changed (and got simplified) quite
a bit in function of the new design.  EPIDs are now allocated as a new
enabling is added to the list (from dt_bpf.c when we are about to load
a program for a specific probe).

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>