Print out any trace data as 64-bit values (temporarily)
Given that the trace() function is currently implemented so that
it is possible to record any number of 64-bit values when a probe
fires, we want to be able to see that data on the userspace level.

Future work will expand upon this to provide the full range of
data output options that DTrace offers.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>