test: allow tests to verify whether a provider is loaded
The DTrace v1 testsuite kept a lsit of loaded providers so that tests
could use the check_provider function to determine if a certain provider
was loaded.  The generation of the list of providers was implemented in
the routine that handled loading the DTrace kernel modules.  Since we no
longer use kernel modules for DTrace v2, the code to generate the provider
list was accidentally removed as well.

This patch brings back code to generate the provider list.  It also fixes
the interpreter specification in one .x file.  It must use the Bourne
Again shell (bash) to have access to the check_provider function.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>