Provide trampoline code for BEGIN, END, and ERROR probes.
THe 'dtrace' provider implements BEGIN, END, and ERROR probes.  The
BEGIN and END probes will be triggered from uprobe events.  This
patch provides the trampoline implementation for these probes.

Note that the ERROR probe is one that gets triggered as a result of
errors occuring during the processing of another probe and that one
will be implemented differently - it is not currently available.

Orabug: 31220513
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>