Implement new dtrace_eprobedesc_t and dtrace_datadesc_t design
In DTrace v1, the dtrace_eprobedesc_t structure was used to obtain
metadata information about the probe data stream from the kernel.
Each EPID was also associated with a specific dtrace_probedesc_t,
identifying the probe for which the EPID represents an enabling.
Each EPID identified a unique dtrace_eprobedesc_t while the
dtrace_probedesc_t could be shared between multiple EPIDs.

    EPID -[1-to-1]-> dtrace_eprobedesc_t
    EPID -[n-to-1]-> dtrace_probedesc_t

In the DTrace v2 design, where everything resides in userspace, we
can create that metadata directly.  We can also make use of the fact
that when a single program is attached to multiple probes, we do not
need more than one copy of the metadata.  We also make use of the fact
that the EPID identifies a unique (probe, program) tuple.

    EPID -[n-to-1]-> dtrace_datadesc_t
    EPID -[n-to-1]-> dtrace_probedesc_t

The probe data description consists of records that each describe one
data item in the probe data.  Records are added as D code is being
compiled, and they specify the type, size, and alignment of the data
item.  The information from the records is used when consuming data
from the probe data buffer.

The dt_rec_add() function is used to register a data item as a new
record in the data description.  It will advance pcb_bufoff to the
proper alignment and call a code-generation callback function to
fill in any needed zero padding.  It returns the offset at which
the data item is to be stored in the output buffer.

This commit also disables the (incomplete) implementation of the
trace() action for string values.  Where before it would simply end
up putting bogus values into the data buffer, it may now cause actual
corruption of the data buffer.

Orabug: 31220520
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>