dlibs: make expand_for_kernel work if invoked more than once
The expand_for_kernel macro had an extremely embarrassing bug:
because it always defined a macro with the same name, and because M4
defines all macros before doing any expansions, calling
define_for_kernel twice always duplicated the expansion of the first
in the second invocation, no matter what the second invocation was.

This is unlikely to be what anyone wants.

Fixed by introducing a name (required to be unique, though we do not
validate this) and using that name to inflect the name of the macro
we define in expand_for_kernel.

Signed-off-by: Nick Alcock <nick.alcock@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>