Provide a standard dt_strtab_copystr() function
Different parts in the DTrace source code were providing their own
custom copy string function to be passed to dt_strtab_write().  All
did essentially the same thing, so it is quite easy to provide a
single generic function to do this work.  It is still possible to
use the dt_strtab_write() functionality with a more fancy copy
string function if needed.

Orabug: 31220525
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>