Introduce -xdisasm=n option
In order to prepare for more functionality for the disassembler in
DTrace an option is being added for specifying which disassembly
listings should be printed if the '-S' option is supplied.

The following listings can be selected:

    No. | Value | Description
    ----+-------+------------
     1  |   1   | After compilation and assembly of a program
     2  |   2   | After linking in precompiled BPF functions
     3  |   4   | After final relocation processing (final program)

The new option is specified as "-xdisasm=n" where n is the bitwise OR of
the values listed in the table above.  The default is 1.

Orabug: 31220524
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>