Implement linking in of precompiled BPF code
Some functionality for D programs is provided as function written in
C and precompiled into BPF instructions (at DTrace build time).  These
functions are available as an ELF object (bpf_dlib.o).  When a compiled
D program requires the use of a precompiled function, that function and
all its dependencies are linked into the program.  When a function is
used more than once, only one copy will be linked in.

The linking of code (implemented as a resolving of relocations that
reference as-yet undefined symbols) takes place right after the D code
has been compiled into a DIFO.  It involves a three-phase approach:

1. We perform a recursive walk of the dependency tree encoded in the
   relocation tables.  The main program lists the functions it depends
   on in the relocation tables, and each function in term lists all of
   its dependencies in its own relocation table.  As we follow the chain
   of dependent functions, we calculate the total number of instructions
   that the final linked program will provide (sum of all unique
   functions), the total number of relocations that the final linked
   program will provide, and the offset of the entry point into each
   function is recorded in the identifier for each of these functions.

2. We again perform a recursive walk of the dependency tree encoded in
   the relocation tables.  This time, we copy the instructions of each
   function, along with its relocations (after adjusting the offset
   based on the new entry address for the function, and setting the
   data to the entry point of the called function).  We also construct
   a combined string table.

3. Finally we process the combined relocation table and patch the code
   accordingly.

A disassembler listing of the linked program can be requested by using
the -xdisasm=2 option along with -S.

An earlier prototype of the BPF code linking process was implemented by
Eugene Loh.

Orabug: 31220525
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>