Fix double-free of dtp->dt_strtab
The consolidated string table that is used to populate the strtab BPF
map when we prepare to load and execute BPF programs was being free'd
twice (first when the DIFOs are destroyed and second when we are closing
the DTrace consumer handle).  The reason for this is that dtp->dt_strtab
was being constructed anew for each DIFO, and then assigned to the
dtdo_strtab member for the DIFO as well.  So, the last DIFO that got
assembled would have a reference to the same string table as the DTrace
consumer handle.

When that DIFO got destroyed (dt_difo_free()), the string table would be
free'd as well.  Then, in dt_close() we would try to free dtp->dt_strtab
which arready got free'd.

We now (correctly) generate string table data for each DIFO, and assign
it to dtp->dt_strtab (size is stored in dtp->strlen).  This is sufficient
because we only care about the last one that get generated (which by
design contains all strings from all prior DIFOs).  We know that the
string table will be free'd in dt_difo_free() so we ensure that the
dtp->dt_strtab and dtp->dt_strlen are cleared when the DIFO referencing
the same string table as the consumer handle is free'd, and we refrain
from freeing dtp->dt_strtab altogether.

Orabug: 31220517
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>