Make flowindent work for general cases
Flowindent was disabled when probe data buffer processing was updated
because there was no support for implementing the heuristic that
requires lookahead for the next probe data sample.  While the heuristic
is still not supported with the current code, there is no need to keep
flowindent disabled because for most cases there is no need to provide
support for epid lookahead.

This patch still allows for a later implementation of lookahead to be
added to dt_consume_cpu() and the value of the next epid is to be passed
to dt_consume_one() and from there to dt_flowindent().

Orabug: 31220522
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>