Introduce internal dt_probe_iter() function
DTrace provides a dtrace_probe_iter() function that can be used by
consumers to iterate over all probes, having a callback function
called for every probe that matches the provided probe description
(which can contain globbing elements).  The callback function is
passed the exact description (id, and fully specified name elements)
for a probe.

For internal purposes callback functions need to pass a pointer to
the actual dt_probe_t structure.

This patch changes the exported dtrace_probe_iter() to simply call
dt_probe_iter() which now provides the actual imeplementation of the
iterator logic.  The dt_probe_id() function accepts two possible
callback functions: one that will be called with a probe description
and one that will be called with a dt_probe_t pointer.  The callback
functions are mutually exclusive, but it is valid to not provide
either.

Orabug: 31220519
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>