Implement probe cleanup function support for providers
Some providers need to be able to do some custom cleanup when a probe
is being destroyed.  This patch adds a callback function to the provider
implementation.  The probe_fini callback will be called for every probe
that belongs to a provider that implements such a callback.

This patch moves the destoying of providers to occur after probes have
been destroyed.  It also changes the allocation of probes to enforce that
the allocated memory is initialized with zeros.  It guards against the
dt_probe_destroy() function being called for dynamically created probes
that are (so far) not linked into the hash tables.

As a little bonus, this patch also fixes one of the argument names in
the prototype of the probe_info callback.

Orabug: 31220519
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>