Make provider name handling in providers more consistent
Providers used a mix of string constants and provname to pass the
provider name in a function call or to name the dt_provimp_t.  Now
prvname (formerly provname) is used anywhere the provider name is
being used.

Other constant elements of probe descriptions (modname and funname for
the dtrace provider) are now also string constants assigned to
variables.

Since provider creation used an assignment within an if-statement and
all these statements were being changed to use prvname, this patch also
separates out the assignment to prv and the if-statement.

Orabug: 31220516
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>