Ensure that both Ctrl-c and exit() terminate probing correctly
Probing can end in two ways: an exit() action or Ctrl-c.  The main
difference is that exit() initiates the completion of probing from
a probe clause (producer) whereas Ctrl-c does so from the probe data
processing (consumer).  In both cases, we want the END probe to fire
(if enabled), and we therefore must ensure that the code path that
trigger it is executed.  This is done by ensuring that when an exit()
action is encountered the consumer ends tracing in a manner similar
it being interrupted (Ctrl-c).

Orabug: 31220520
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>