Implement BEGIN and END probes for the dtrace provider
Add BEGIN and END probes as uprobes on libdtrace functions BEGIN_probe()
and END_probe(), which were added as part of an earlier patch.

By default, uprobes belong to group "uprobes", but we introduce a new
group, "dtrace_$pid", with probes that are specific to our consumer
process.  Specifically, this means that multiple consumers, with
distinct or changing libdtrace, can coexist.  We try to clean up our
BEGIN and END probes at close, and the sdt provider must recognize that
dtrace_$pid is a group to ignore.

Orabug: 31220513
Signed-off-by: Eugene Loh <eugene.loh@oracle.com>
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>