Fix value of post-increment expressions
When using a post-increment (or post-decrement) on a variable of any
kind, the expression should evaluate to the old value.

E.g. trace(this->x++) generates (WRONG):

	lddw	%r8, [%fp-80]		! load old value
	add	%r8, 1			! increment
	stdw	[%fp-80], %r8		! store new value
	(%r8 is used as value, which is old value + 1)

while instead it should generate (CORRECT):

	lddw	%r8, [%fp-80]		! load old value
	mov	%r7, 1			! load increment
	add	%r7, %r8		! add old value to increment
	stdw	[%fp-80], %r7		! store new value
	(%r8 is used as value, which is old value)

Tests are included in this patch for both pre-increment and
post-increment of each of the three variable kinds.  Since the code
generator uses the same logic for increment and decrement, testing
just one is sufficient.

Orabug: 31220528
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>