Check register allocation requests and handle function calls
Some register allocation calls did not check the return value to ensure
we were able to allocate the register.  That has been fixed.

BPF function calls (to BPF functions or helper functions) potentially
clobber %r1 through %r5.  The verifier ensures that these registers
cannot be read from after a call until an explicit store is done into
the register.

This patch adds dt_regset_xalloc_args() and dt_regset_free_args() as
convenient ways to bulk allocate and free %r1 through %r5.  The alloc
fails if any of the registers is already in use.

Due to the reservation of %r9 as output buffer base pointer, the need
for %r1 through %r5 as function arguments, and the use of %r0 as call
return value (even if none is returned), only %r6 through %r8 are
available as general scratch registers in code that makes function
calls.  This imposes a severe limitation.  In order to provide for an
adequate number of general purpose registers, register spilling is
implemented.

Orabug: 31187562
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>