Implementation of the exit() action
The exit() action is quite instrumental to the use of DTrace because it
allows a tracing script to terminate the tracing session.

Because exit() must be able to signal that trace data processing is
done, dt_consume_one(), dt_consume_cpu(), and dt_consume() have been
updated to return a dtrace_workstatus_t value to indicate whether
processing of trace data is to continue (DTRACE_WORKSTATUS_OKAY),
is done (DTRACE_WORKSTATUS_DONE), or whether an error occurred
(DTRACE_WORKSTATUS_ERROR).

Orabug: 31220520
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Kris Van Hees <kris.van.hees@oracle.com>
Reviewed-by: Eugene Loh <eugene.loh@oracle.com>