Fix the stack layout for register spilling
The layout of the stack and the constants controlling that were poorly
defined at the onset of the DTrace v2 development.  As a result, when
register spilling was implemented, various bugs came to light.  E.g. it
was possible to overwrite other fields in the stack (cpu, local
variables).

This patch includes tests that verify that the stack layout is correct,
although it should be noted that the stack corruption can happen in
more subtle ways and not all cases can be tested for in a formal manner.
The tests should cover the more comon cases though.

This patch also adds a test utility to print out the offset of the
various fields in the stack used by DTrace for BPF programs (and one of
the tests uses this).

Since this patch makes some modifications to the build system machinery
for installing the print-stack-layout utlity, a little additional
reworking of the install-test target for utils is included to ensure
that executables are installed correctly.  It also removes the ETCDIR
and INSTETCDIR makefile variables that are no longer used.

Orabug: 31187562
Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>