Temporary fix to avoid excessive stack usage
For kernels that have CONFIG_BPF_JIT_ALWAYS_ON=y the BPF verifier checks
that the total stack use of a call chain (sequence of function calls)
is not greater than the maximum BPF stack size (currently set at 512
bytes in the kernel).

The trampoline code constructs the DTrace BPF context on the stack, and
then calls dt_predicate and dt_program.  The dt_program always starts
with a common prologue that stores some information on the stack.  We
also use stack locations for register spilling.  Finally, various helper
functions that are linked into our program may also make use of stack
slots.

The DTrace BPF context is the largest static structure on the stack and
it is architecture dependent due to the inclusion of pt_regs.  This
poses a problem on arm64 where the pt_regs structure is significantly
larger than on x86_64.

A new stack layout design is in the works which will make it possible to
rework this temporary fix.

Signed-off-by: Kris Van Hees <kris.van.hees@oracle.com>